WasteNoughtBE/requirements.txt
alembic==1.13.3
asttokens==2.4.1
autopep8==2.0.4
bcrypt==4.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
executing==2.0.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-DebugToolbar==0.14.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
iniconfig==2.0.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
Mako==1.3.5
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
numpy==1.26.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pluggy==1.4.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
Pygments==2.17.2
PyJWT==2.9.0
pytest==7.4.4
python-dotenv==1.0.1
requests==2.31.0
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
WTForms==3.1.2
